+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 See `UPGRADE.md` for instructions on upgrading between versions.
 
+## 0.44.0
+
+* Rename `Brando.Field.ImageField` -> `Brando.Field.Image.Schema`
+* Rename `Brando.Field.FileField` -> `Brando.Field.File.Schema`
+
+
 ## 0.43.0
 
 * Add `Brando.HTML.init_js()`